Fixes maps tooltips!
THIS COMMIT UPDATES A VENDOR FILE. See below for details.

I investigated on why these tooltips weren't showing. It all started
when I noticed tooltips were actually displayed… but for the latest
map only! I had the hypothesis there was a bug with multiple maps
in the same page and tooltips.

I dug into the source code and found that indeed, when the tip
element is created, other similar elements are removed without
clear reason behind that. As only a reference to the tooltip element
is kept afterwards, the Element is actually updated, but as the
corresponding DOM element no longer exists (except for the latest
initialized map!), nothing is visible.

I patched the jvectormap to remove the removal of other tips and it
fixes the bug!

Also I moved the `onRegionTipShow` to the root of the parameters to
be able to access original region code and with that, the signatures
count for this department.

And I deleted the compressed version of jvectormap's JS/CSS source,
to avoid confusion, as these two no longer match.

Some pull request should be made upstream—this is a hacky quick fix!